requirements.txt
discord-interactions>=0.2.0
google-api-python-client==2.5.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-pubsub==2.5.0
requests==2.25.1